+++ CHANGELOG.md
@@ -137,9 +137,9 @@ Allow Laravel 9
 ## 2.0.0 - 2016-06-20
 
 - refactored all methods
-- introduced `Spatie\Analytics\Period` to specify date ranges
+- introduced `Spatie\Universal\Analytics\Period` to specify date ranges
 - the package now uses json credentials instead of .p12 file
-- the `Spatie\Analytics\Analtyics` class is much easier to extend
+- the `Spatie\Universal\Analytics\Analtyics` class is much easier to extend
 
 ## 1.4.1 - 2016-05-12